--- app/modules/annotations/annotations.config.js
@@ -1,9 +1,0 @@
-'use strict';
-
-var annotationsConfig = {
-    pointRadius: 10
-};
-
-module.exports = {
-    constants: annotationsConfig
-};
