--- app/modules/annotations/image.html
@@ -1,10 +1,0 @@
-<g class="image-annotation">
-    <rect
-        class="image"
-        draggable="data"
-        ng-attr-x="{{ data.x }}"
-        ng-attr-y="{{ data.y }}"
-        ng-attr-height="{{ data.height }}"
-        ng-attr-width="{{ data.width }}"
-    ></rect>
-</g>
