--- app/modules/marking-tools/marking-tools.module.js
@@ -1,5 +1,0 @@
-'use strict';
-
-// A module for the static pages such as home, about etc.
-module.exports = require('angular')
-    .module('app.markingTools', []);
