--- app/modules/overlay/context-menu.directive.js
@@ -1,100 +1,0 @@
-'use strict';
-
-var _ = require('lodash');
-var angular = require('angular');
-var Hammer = require('hammerjs');
-
-require('./overlay.module.js')
-    .directive('contextMenu', contextMenu);
-
-// TODO: Add escape hotkey to close
-// TODO: Add arrow key / spacebar support
-// TODO: Find out what ngInject isn't working properly for contextMenuController
-
-// @ngInject
-function contextMenu($rootScope, $window, hotkeys) {
-    var directive = {
-        controller: ['$scope', contextMenuController],
-        link: contextMenuLink,
-        replace: true,
-        scope: true,
-        templateUrl: 'overlay/context-menu.html'
-    };
-    return directive;
-
-    // @ngInject
-    function contextMenuController($scope) {
-
-        // Setup
-        var vm = this;
-        vm.close = closeMenu;
-        vm.open = openMenu;
-
-        // Methods
-        function closeMenu() {
-            $scope.active = false;
-            hotkeys.del('esc');
-        }
-
-        function openMenu(data) {
-            $scope.active = true;
-            $scope.menuOptions = data.menuOptions;
-            hotkeys.add({
-                combo: 'esc',
-                callback: closeMenu
-            });
-        }
-    }
-
-    // @ngInject
-    function contextMenuLink(scope, element, attrs, contextMenu) {
-
-        // Setup
-        var bodyEvent;
-        var overlay;
-        overlay = angular.element('.overlay').first();
-        scope.position = {};
-
-        // Events
-        scope.$on('contextMenu:open', openContextMenu);
-
-        // There's a bug in the current version of Hammer preventing event
-        // binding to the window object, so we use body as a workaround.
-        // https://github.com/hammerjs/hammer.js/issues/759
-        bodyEvent = new Hammer(element.parents('body')[0]);
-
-        // Methods
-
-        function closeContextMenu() {
-            $rootScope.$broadcast('panZoom:enable');
-            $rootScope.$broadcast('markingTools:enable');
-            contextMenu.close();
-            scope.$apply();
-            bodyEvent.off('tap', this);
-        }
-        function openContextMenu(event, data) {
-            $rootScope.$broadcast('panZoom:disable');
-            $rootScope.$broadcast('markingTools:disable');
-            contextMenu.open(data);
-            positionContextMenu(data);
-            bodyEvent.on('tap', closeContextMenu);
-        }
-
-        function positionContextMenu(data) {
-            var click = data.event.srcEvent;
-            scope.position = {
-                left: click.offsetX,
-                top: click.offsetY
-            };
-
-            // Firefox doesn't support offset, so we need to polyfill here.
-            if (_.isUndefined(click.offsetX) || _.isUndefined(click.offsetY)) {
-                scope.position = {
-                    left: click.pageX - overlay.offset().left,
-                    top: click.pageY - overlay.offset().top
-                };
-            }
-            scope.$apply();
-        }
-    }
-}
