--- app/modules/overlay/context-menu.html
@@ -1,13 +1,0 @@
-<ul
-    aria-hidden="{{ !active }}"
-    class="context-menu"
-    ng-class="{ '-active': active }"
-    ng-style="position"
-    role="menu"
-    >
-
-    <li ng-repeat="option in menuOptions">
-        <a href ng-click="option.action()" ng-bind="option.name"></a>
-    </li>
-
-</ul>
