--- app/modules/overlay/loading-spinner.directive.js
@@ -1,22 +1,0 @@
-'use strict';
-
-var Spinner = require('spin');
-
-require('./overlay.module.js')
-    .directive('loadingSpinner', loadingSpinner);
-
-// @ngInject
-function loadingSpinner() {
-    var directive = {
-        compile: loadingSpinnerCompile,
-        replace: true,
-        templateUrl: 'overlay/loading-spinner.html'
-    };
-    return directive;
-
-    function loadingSpinnerCompile(tElement) {
-        new Spinner({
-            color: '#fff'
-        }).spin(tElement[0]);
-    }
-}
