--- app/modules/overlay/loading-spinner.html
@@ -1,1 +1,0 @@
-<div class="loading-spinner" ng-hide="vm.subject.isLoaded"></div>
