--- app/modules/overlay/out-of-data.html
@@ -1,14 +1,0 @@
-<div
-    ng-show="active"
-    class="outofdata-dialog"
-    >
-
-    <div class="heading">
-        <h3 class="title">Out of data</h3>
-    </div>
-
-    <div class="body">
-        <p>No more data available in this set.</p>
-    </div>
-
-</div>
