--- app/modules/overlay/overlay.html
@@ -1,6 +1,0 @@
-<div class="overlay-layer">
-    <div loading-spinner></div>
-    <div context-menu></div>
-    <div transcribe-dialog></div>
-    <div out-of-data></div>
-</div>
