--- app/modules/overlay/overlay.module.js
@@ -1,7 +1,0 @@
-'use strict';
-
-require('spin');
-
-// A module for context menus, editing windows etc
-module.exports = require('angular')
-    .module('app.overlay', []);
