--- app/modules/subjects/ngheight.directive.js
@@ -1,18 +1,0 @@
-'use strict';
-
-require('./subjects.module.js')
-    .directive('ngHeight', ngHeight);
-
-// @ngInject
-function ngHeight() {
-    var directive = {
-        link: linkFunction
-    };
-    return directive;
-
-    function linkFunction(scope, element, attrs) {
-        scope.$watch(attrs.ngHeight, function (value) {
-            element.attr('height', value);
-        });
-    }
-}
