--- app/modules/subjects/subjects.module.js
@@ -1,8 +1,0 @@
-'use strict';
-
-require('ng-load');
-
-module.exports = require('angular')
-    .module('app.subjects', [
-        'ngLoad'
-    ]);
